pyproject.toml
[project]
name = "skeeball-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-aiplatform",
    "google-generativeai",
    "google-cloud-pubsub",
    "google-cloud-bigquery",
    "pyyaml"  # Add PyYAML dependency
]
